minor modification. 6.3 seconds for 500MB data

+++ algorithm.h
@@ -32,10 +32,11 @@
 
 using namespace std;
 
-void get_anno_info(ifstream& in_anno, const int anno_choice, map<string, gene_info> & map_g_anno);
+void get_anno_info(ifstream& in_anno, const unsigned int anno_choice,
+    map<string, gene_info> & map_g_anno);
 
 // get the read count of each exon
-int get_exon_rd_cnt(map<string, gene_info> & map_g_info, ifstream & in_rdmap,
+size_t get_exon_rd_cnt(map<string, gene_info> & map_g_info, ifstream & in_rdmap,
     size_t & tot_valid_rd_cnt, vector<double> & GBC);
 
 void calcuAllTheGenes(map<string, gene_info> & map_g_info,
